+++ src/main/java/dinkplugin/notifiers/CollectionNotifier.java
@@ -2,11 +2,15 @@ package dinkplugin.notifiers;
 
 import dinkplugin.message.NotificationBody;
 import dinkplugin.message.NotificationType;
+import dinkplugin.util.ItemSearcher;
+import dinkplugin.util.ItemUtils;
 import dinkplugin.util.Utils;
 import dinkplugin.notifiers.data.CollectionNotificationData;
+import net.runelite.client.game.ItemManager;
 import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.VisibleForTesting;
 
+import javax.inject.Inject;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -14,6 +18,12 @@ public class CollectionNotifier extends BaseNotifier {
     @VisibleForTesting
     static final Pattern COLLECTION_LOG_REGEX = Pattern.compile("New item added to your collection log: (?<itemName>(.*))");
     public static final String ADDITION_WARNING = "Collection notifier will not fire unless you enable the game setting: Collection log - New addition notification";
+
+    @Inject
+    private ItemManager itemManager;
+
+    @Inject
+    private ItemSearcher itemSearcher;
 
     @Override
     public boolean isEnabled() {
@@ -41,9 +51,12 @@ public class CollectionNotifier extends BaseNotifier {
             new String[] { Utils.getPlayerName(client), itemName }
         );
 
+        Integer itemId = itemSearcher.findItemId(itemName);
+        Long price = itemId != null ? ItemUtils.getPrice(itemManager, itemId) : null;
         createMessage(config.collectionSendImage(), NotificationBody.builder()
             .text(notifyMessage)
-            .extra(new CollectionNotificationData(itemName))
+            .thumbnailUrl(itemId != null ? ItemUtils.getItemImageUrl(itemId) : null)
+            .extra(new CollectionNotificationData(itemName, itemId, price))
             .type(NotificationType.COLLECTION)
             .build());
     }