+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Minor: Reduce ticks to initialize diary notifier and improve logging for edge cases. (#129)
+- Minor: Made Collection Log notifications display the icon of the obtained item. (#128)
 - Minor: Made player names in Discord notifications link to a player profile page. This feature can be customized to different providers such as WiseOldMan & CrystalMathLabs in the Advanced menu. (#126)
 - Minor: Add setting to skip virtual level notifications. (#122)
 - Minor: Update the README to be less confusing to users. (#123)